# quarkus-astro-app/src/main/webui/api/openapi.yaml
...
openapi: 3.1.0
...
components:
  schemas:
...
    CategoryResponseDTO:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        colorCode:
          type: string
        isDefault:
          type: boolean
        createdAt:
          $ref: "#/components/schemas/LocalDateTime"
        taskCount:
          type: integer
          format: int64
...
    LocalDateTime:
      type: string
      format: date-time
      examples:
      - 2022-03-10T12:15:50
    Priority:
      type: string
      enum:
      - HIGH
      - MEDIUM
      - LOW
    TaskCreateDTO:
      type: object
      required:
      - title
      - categoryId
      - priority
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        categoryId:
          $ref: "#/components/schemas/UUID"
        priority:
          $ref: "#/components/schemas/Priority"
    TaskResponseDTO:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        title:
          type: string
        description:
          type: string
        category:
          $ref: "#/components/schemas/CategoryResponseDTO"
        priority:
          $ref: "#/components/schemas/Priority"
        completed:
          type: boolean
        completedAt:
          $ref: "#/components/schemas/LocalDateTime"
        createdAt:
          $ref: "#/components/schemas/LocalDateTime"
        updatedAt:
          $ref: "#/components/schemas/LocalDateTime"
    TaskUpdateDTO:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        categoryId:
          $ref: "#/components/schemas/UUID"
        priority:
          $ref: "#/components/schemas/Priority"
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
...
paths:
...
  /api/tasks:
    get:
      summary: List user's tasks
      description: "Retrieve tasks with optional filtering by category, priority,\
        \ and status"
      tags:
      - Tasks
      parameters:
      - description: Filter by category UUID
        in: query
        schema:
          format: uuid
          type: string
        name: category
      - description: Page number (0-based)
        in: query
        schema:
          minimum: 0
          type: integer
          default: 0
        name: page
      - description: Filter by priority level
        in: query
        schema:
          $ref: "#/components/schemas/Priority"
        name: priority
      - description: Page size
        in: query
        schema:
          maximum: 100
          minimum: 1
          type: integer
          default: 20
        name: size
      - description: Filter by completion status
        in: query
        schema:
          type: string
          enum:
          - active
          - completed
        name: status
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponseDTO"
        "401":
          description: Not authenticated
    post:
      summary: Create a new task
      description: Creates a new task for the authenticated user
      tags:
      - Tasks
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreateDTO"
        required: true
      responses:
        "201":
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponseDTO"
        "400":
          description: Invalid request body or validation error
        "401":
          description: Not authenticated
  /api/tasks/{id}:
    put:
      summary: Update a task
      description: Updates an existing task. All fields in the request body are optional.
      tags:
      - Tasks
      parameters:
      - description: Task UUID
        required: true
        schema:
          format: uuid
          type: string
        name: id
        in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdateDTO"
        required: true
      responses:
        "200":
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponseDTO"
        "400":
          description: Invalid request body or validation error
        "404":
          description: Task not found
        "401":
          description: Not authenticated
    get:
      summary: Get task by ID
      description: Retrieves a single task by its unique identifier
      tags:
      - Tasks
      parameters:
      - description: Task UUID
        required: true
        schema:
          format: uuid
          type: string
        name: id
        in: path
      responses:
        "200":
          description: Task found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponseDTO"
        "404":
          description: Task not found
        "401":
          description: Not authenticated
    delete:
      summary: Delete a task
      description: Deletes an existing task
      tags:
      - Tasks
      parameters:
      - description: Task UUID
        required: true
        schema:
          format: uuid
          type: string
        name: id
        in: path
      responses:
        "204":
          description: Task deleted successfully
        "404":
          description: Task not found
        "401":
          description: Not authenticated
  /api/tasks/{id}/complete:
    patch:
      summary: Toggle task completion
      description: "Toggles a task between complete and incomplete. If incomplete,\
        \ marks as complete with timestamp. If complete, marks as incomplete and clears\
        \ timestamp."
      tags:
      - Tasks
      parameters:
      - description: Task UUID
        required: true
        schema:
          format: uuid
          type: string
        name: id
        in: path
      responses:
        "200":
          description: Task completion toggled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponseDTO"
        "404":
          description: Task not found
        "401":
          description: Not authenticated
info:
  title: Task Manager API
  version: 1.0.0
  description: RESTful API for Task Manager application demonstrating Islands Architecture
  contact:
    name: Task Manager Team
servers:
- url: http://localhost:7171
  description: Auto generated value
- url: http://0.0.0.0:7171
  description: Auto generated value