# quarkus-astro-app/src/main/webui/src/lib/api/model/taskUpdateDTO.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Task Manager API
 * RESTful API for Task Manager application demonstrating Islands Architecture
 * OpenAPI spec version: 1.0.0
 */
import type { Uuid } from './uuid';
import type { Priority } from './priority';

export interface TaskUpdateDTO {
  /**
   * @minLength 1
   * @maxLength 200
   */
  title?: string;
  /** @maxLength 2000 */
  description?: string;
  categoryId?: Uuid;
  priority?: Priority;
}